Calculate Gap Statistics for different time ranges

Given a univariate daily time series, the following statistics are computed for the specified date ranges:

Completeness Percentage, daily (c_d): Percentage of non-NA values in the time series

Completeness Percentage, monthly (c_m): Percentage of non-NA values after aggregating to monthly

Longest Gap,monthly (l_g): Longest consecutive stretch of NA values after aggregating to monthly

gap_statistics(x, date_ranges, completeness_threshold = 100,
  md = cl_catchment_md)

Arguments

x
daily timeseries as tibble with Date and data columns named with the stationcode

date_ranges
character vector of date ranges

completeness_threshold
completeness percentage below which a month is marked as NA when calculating monthly statistics (c_m, l_g)

md
metadata table (tibble) with stationcode column

Value

the metadata table md with the following vars added: stationcode + c_d, c_m, l_g for each date range

Details

The time ranges are specified in years as "start-end", with "2000-2009" meaning from 2000-01-01 to 2009-12-31.